Refactor the huggingface model code
